Migrate transaction pages from Next to React Router
