+++ src/main.tsx
@@ -1,45 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
- codex/implement-usetheme-hook-and-context
-import App from './App';
+import { RouterProvider } from 'react-router-dom';
 import './index.css';
+import { router } from './router';
 import { ThemeProvider } from './theme';
+import { FinanceProvider } from './store/FinanceContext';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <ThemeProvider>
-      <App />
+      <FinanceProvider>
+        <RouterProvider router={router} />
+      </FinanceProvider>
     </ThemeProvider>
-
- codex/create-kpi-cards-with-recharts
-
- codex/create-dashboardlayout-components-and-routing
- dev
-import App from './App';
-import './index.css';
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <App />
- codex/create-kpi-cards-with-recharts
-
-
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import App from './App.tsx';
-import './index.css';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-  },
-]);
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <RouterProvider router={router} />
- dev
- dev
- dev
   </React.StrictMode>
 );
