--- src/pages/_app.tsx
@@ -1,11 +1,0 @@
-import type { AppProps } from 'next/app';
-import { FinanceProvider } from '../store/FinanceContext';
-import '../styles/globals.css';
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <FinanceProvider>
-      <Component {...pageProps} />
-    </FinanceProvider>
-  );
-}
