+++ tsconfig.node.json
@@ -1,36 +1,14 @@
 {
- codex/implement-usetheme-hook-and-context
-
- codex/create-kpi-cards-with-recharts
-
- codex/create-dashboardlayout-components-and-routing
- dev
- dev
-  "compilerOptions": {
-    "composite": true,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["vite.config.ts"]
- codex/implement-usetheme-hook-and-context
-
- codex/create-kpi-cards-with-recharts
-
-
   "$schema": "https://json.schemastore.org/tsconfig",
   "compilerOptions": {
     "composite": true,
-    "skipLibCheck": true,
     "module": "ESNext",
     "moduleResolution": "Node",
     "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
     "types": ["node"]
   },
-  "include": ["vite.config.ts", "tsconfig.node.json"]
- dev
- dev
- dev
+  "include": ["vite.config.ts"]
 }
